(#13) refactor: change send message
